+++ spring/src/main/java/br/net/yurinogueira/springsales/rest/dto/SaleInfoDTO.java
@@ -1,7 +1,5 @@
 package br.net.yurinogueira.springsales.rest.dto;
 
-import java.math.BigDecimal;
-
 import br.net.yurinogueira.springsales.domain.enums.SaleType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -25,6 +23,6 @@ public class SaleInfoDTO {
 
     private Integer saleAmount;
 
-    private BigDecimal salePrice;
+    private Double salePrice;
 
 }
